[src/Write] make "stream indices" of same type as in "struct BlockHeader"
class StreamWrite and Class Output:

change member-variables and return-values of functions to have "uint16_t" as their value will beu sed to fil the "BlockHeader" later.

+++ src/Write.cpp
@@ -64,7 +64,7 @@ namespace pocolog_cpp
         writePrologue(stream);
     }
 
-    int Output::newStreamIndex()
+    uint16_t Output::newStreamIndex()
     { return m_stream_idx++; }
 
     void Output::writeStreamDeclaration(uint16_t stream_index, StreamType type,
