+++ src/Write.hpp
@@ -52,7 +52,7 @@ namespace pocolog_cpp
         static const size_t nstream = static_cast<size_t>(-1);
 
         std::ostream& m_stream;
-        int m_stream_idx;
+        uint16_t m_stream_idx;
 
     private:
         template<class T>
@@ -68,7 +68,7 @@ namespace pocolog_cpp
 
         std::ostream& getStream();
 
-        int newStreamIndex();
+        uint16_t newStreamIndex();
 
         void writeStreamDeclaration(uint16_t stream_index, StreamType type,
                 std::string const& name, std::string const& type_name,
@@ -167,8 +167,9 @@ namespace pocolog_cpp
         std::string const m_type_name;
         std::string const m_type_def;
         std::vector<StreamMetadata> m_metadata;
-        int const m_stream_idx;
-        size_t const m_type_size;
+        // same types as in "struct BlockHeader"
+        uint16_t const m_stream_idx;
+        uint32_t const m_type_size;
         base::Time m_sampling;
         base::Time m_last;
